povorotnaya.png:replace map preview (fixes #30)
Before this change, I wasn't having any problems testing using the git
version of cm2; however, when the pyromod is created, the png files are
converted and wind up with a '.cached.dds' extension. I did not test the
mod using the pyromod file before releasing v0.25.10

Thanks to @N0mi5 for the bug report.